fix: load ctx resource only once
