+++ cmd/context/utils.go
@@ -26,6 +26,7 @@ import (
 	"github.com/okteto/okteto/pkg/k8s/kubeconfig"
 	oktetoLog "github.com/okteto/okteto/pkg/log"
 	"github.com/okteto/okteto/pkg/model"
+	"github.com/okteto/okteto/pkg/model/contextresource"
 	"github.com/okteto/okteto/pkg/okteto"
 	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
 )
@@ -112,7 +113,7 @@ func isValidCluster(cluster string) bool {
 	return false
 }
 
-func addKubernetesContext(cfg *clientcmdapi.Config, ctxResource *model.ContextResource) error {
+func addKubernetesContext(cfg *clientcmdapi.Config, ctxResource *contextresource.ContextResource) error {
 	if cfg == nil {
 		return fmt.Errorf(oktetoErrors.ErrKubernetesContextNotFound, ctxResource.Context, config.GetKubeconfigPath())
 	}
@@ -129,21 +130,12 @@ func addKubernetesContext(cfg *clientcmdapi.Config, ctxResource *model.ContextRe
 	return nil
 }
 
+// LoadManifestWithContext loads context and then loads a manifest
 func LoadManifestWithContext(ctx context.Context, opts ManifestOptions) (*model.Manifest, error) {
-
-	var manifest *model.Manifest
-	manifest, err := model.GetManifestV1(opts.Filename)
+	ctxResource, err := contextresource.Get(opts.Filename)
 	if err != nil {
-		if !errors.Is(err, oktetoErrors.ErrManifestNotFound) {
-			return nil, err
-		}
-		manifest, err = model.GetManifestV2(opts.Filename)
-		if err != nil {
-			return nil, err
-		}
+		return nil, err
 	}
-
-	ctxResource := model.GetContextResourceFromManifest(manifest)
 	if err := ctxResource.UpdateNamespace(opts.Namespace); err != nil {
 		return nil, err
 	}
@@ -163,13 +155,12 @@ func LoadManifestWithContext(ctx context.Context, opts ManifestOptions) (*model.
 	}
 
 	// We need to read it again to propagate secrets env vars
-	if manifest.IsV2 {
-		manifest, err = model.GetManifestV2(opts.Filename)
-		if err != nil {
+	manifest, err := model.GetManifestV1(opts.Filename)
+	if err != nil {
+		if !errors.Is(err, oktetoErrors.ErrManifestNotFound) {
 			return nil, err
 		}
-	} else {
-		manifest, err = model.GetManifestV1(opts.Filename)
+		manifest, err = model.GetManifestV2(opts.Filename)
 		if err != nil {
 			return nil, err
 		}
@@ -196,7 +187,7 @@ func LoadStackWithContext(ctx context.Context, name, namespace string, stackPath
 		if name == "" {
 			return nil, err
 		}
-		ctxResource = &model.ContextResource{}
+		ctxResource = contextresource.NewContextResource()
 	}
 
 	if err := ctxResource.UpdateNamespace(namespace); err != nil {
@@ -224,34 +215,31 @@ func LoadStackWithContext(ctx context.Context, name, namespace string, stackPath
 	return s, nil
 }
 
-//LoadManifestV2WithContext initializes the okteto context taking into account command flags and manifest namespace/context fields
-func LoadManifestV2WithContext(ctx context.Context, namespace, k8sContext, path string) error {
+//LoadContextFromPath initializes the okteto context taking into account command flags and manifest namespace/context fields
+func LoadContextFromPath(ctx context.Context, namespace, k8sContext, path string) error {
 	ctxOptions := &ContextOptions{
 		Namespace: namespace,
 		Show:      true,
 	}
-
-	manifest, err := model.GetManifestV2(path)
+	ctxResource, err := contextresource.Get(path)
 	if err != nil {
-		if err != oktetoErrors.ErrManifestNotFound {
+		if !errors.Is(err, oktetoErrors.ErrManifestNotFound) {
 			return err
 		}
-	} else {
-		ctxResource := model.GetContextResourceFromManifest(manifest)
+	}
 
-		if err := ctxResource.UpdateNamespace(namespace); err != nil {
-			return err
-		}
+	if err := ctxResource.UpdateNamespace(namespace); err != nil {
+		return err
+	}
 
-		if err := ctxResource.UpdateContext(k8sContext); err != nil {
-			return err
-		}
+	if err := ctxResource.UpdateContext(k8sContext); err != nil {
+		return err
+	}
 
-		ctxOptions = &ContextOptions{
-			Context:   ctxResource.Context,
-			Namespace: ctxResource.Namespace,
-			Show:      true,
-		}
+	ctxOptions = &ContextOptions{
+		Context:   ctxResource.Context,
+		Namespace: ctxResource.Namespace,
+		Show:      true,
 	}
 
 	return NewContextCommand().Run(ctx, ctxOptions)
