+++ cmd/context/utils_test.go
@@ -19,6 +19,7 @@ import (
 	"testing"
 
 	"github.com/okteto/okteto/pkg/model"
+	"github.com/okteto/okteto/pkg/model/contextresource"
 	"github.com/okteto/okteto/pkg/okteto"
 	"github.com/stretchr/testify/assert"
 	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
@@ -28,14 +29,14 @@ func Test_addKubernetesContext(t *testing.T) {
 	var tests = []struct {
 		name         string
 		cfg          *clientcmdapi.Config
-		ctxResource  *model.ContextResource
+		ctxResource  *contextresource.ContextResource
 		currentStore *okteto.OktetoContextStore
 		wantStore    *okteto.OktetoContextStore
 		wantError    bool
 	}{
 		{
 			name:        "nil-cfg",
-			ctxResource: &model.ContextResource{Context: "context"},
+			ctxResource: &contextresource.ContextResource{Context: "context"},
 			wantError:   true,
 		},
 		{
@@ -43,7 +44,7 @@ func Test_addKubernetesContext(t *testing.T) {
 			cfg: &clientcmdapi.Config{
 				Contexts: map[string]*clientcmdapi.Context{},
 			},
-			ctxResource: &model.ContextResource{Context: "context"},
+			ctxResource: &contextresource.ContextResource{Context: "context"},
 			wantError:   true,
 		},
 		{
@@ -51,7 +52,7 @@ func Test_addKubernetesContext(t *testing.T) {
 			cfg: &clientcmdapi.Config{
 				Contexts: map[string]*clientcmdapi.Context{"context": {Namespace: "n-cfg"}},
 			},
-			ctxResource: &model.ContextResource{Context: "context", Namespace: "n-ctx"},
+			ctxResource: &contextresource.ContextResource{Context: "context", Namespace: "n-ctx"},
 			currentStore: &okteto.OktetoContextStore{
 				CurrentContext: "",
 				Contexts:       map[string]*okteto.OktetoContext{},
@@ -69,7 +70,7 @@ func Test_addKubernetesContext(t *testing.T) {
 			cfg: &clientcmdapi.Config{
 				Contexts: map[string]*clientcmdapi.Context{"context": {Namespace: "n-cfg"}},
 			},
-			ctxResource: &model.ContextResource{Context: "context"},
+			ctxResource: &contextresource.ContextResource{Context: "context"},
 			currentStore: &okteto.OktetoContextStore{
 				CurrentContext: "",
 				Contexts:       map[string]*okteto.OktetoContext{},
@@ -87,7 +88,7 @@ func Test_addKubernetesContext(t *testing.T) {
 			cfg: &clientcmdapi.Config{
 				Contexts: map[string]*clientcmdapi.Context{"context": {}},
 			},
-			ctxResource: &model.ContextResource{Context: "context"},
+			ctxResource: &contextresource.ContextResource{Context: "context"},
 			currentStore: &okteto.OktetoContextStore{
 				CurrentContext: "",
 				Contexts:       map[string]*okteto.OktetoContext{},
