+++ cmd/deploy/endpoints.go
@@ -26,6 +26,7 @@ import (
 	oktetoErrors "github.com/okteto/okteto/pkg/errors"
 	oktetoLog "github.com/okteto/okteto/pkg/log"
 	"github.com/okteto/okteto/pkg/model"
+	"github.com/okteto/okteto/pkg/model/contextresource"
 	"github.com/okteto/okteto/pkg/okteto"
 	"github.com/spf13/cobra"
 )
@@ -57,7 +58,7 @@ func Endpoints(ctx context.Context) *cobra.Command {
 			ctxResource, err := utils.LoadManifestContext(options.ManifestPath)
 			if err != nil {
 				if oktetoErrors.IsNotExist(err) {
-					ctxResource = &model.ContextResource{}
+					ctxResource = &contextresource.ContextResource{}
 				}
 			}
