--- pkg/model/context.go
@@ -1,65 +1,0 @@
-package model
-
-import (
-	"os"
-
-	oktetoErrors "github.com/okteto/okteto/pkg/errors"
-	yaml "gopkg.in/yaml.v2"
-)
-
-type ContextResource struct {
-	Context   string
-	Namespace string
-}
-
-// GetContextResourceFromManifest retrieves the ctxresource from a manifest
-func GetContextResourceFromManifest(m *Manifest) *ContextResource {
-	ctxResource := &ContextResource{}
-	if m != nil {
-		ctxResource.Context = os.ExpandEnv(m.Context)
-		ctxResource.Namespace = os.ExpandEnv(m.Namespace)
-	}
-	return ctxResource
-}
-
-// GetContextResource returns a ContextResource object from a given file
-func GetContextResource(filePath string) (*ContextResource, error) {
-	bytes, err := os.ReadFile(filePath)
-	if err != nil {
-		return nil, err
-	}
-
-	ctxResource := &ContextResource{}
-	if err := yaml.Unmarshal(bytes, ctxResource); err != nil {
-		return nil, err
-	}
-
-	ctxResource.Context = os.ExpandEnv(ctxResource.Context)
-	ctxResource.Namespace = os.ExpandEnv(ctxResource.Namespace)
-
-	return ctxResource, nil
-}
-
-func (c *ContextResource) UpdateContext(okCtx string) error {
-	if c.Context != "" {
-		if okCtx != "" && okCtx != c.Context {
-			return oktetoErrors.ErrContextNotMatching
-		}
-		return nil
-	}
-
-	c.Context = okCtx
-	return nil
-}
-
-func (c *ContextResource) UpdateNamespace(okNs string) error {
-	if c.Namespace != "" {
-		if okNs != "" && c.Namespace != okNs {
-			return oktetoErrors.ErrNamespaceNotMatching
-		}
-		return nil
-	}
-
-	c.Namespace = okNs
-	return nil
-}
