+++ pkg/model/contextresource/contextresource_test.go
@@ -1,4 +1,16 @@
-package model
+// Copyright 2022 The Okteto Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+package contextresource
 
 import (
 	"os"
@@ -45,7 +57,7 @@ func Test_GetContextResource(t *testing.T) {
 			for k, v := range tt.env {
 				os.Setenv(k, v)
 			}
-			result, err := GetContextResource(tmpFile.Name())
+			result, err := Get(tmpFile.Name())
 			if err != nil {
 				t.Fatalf("error reading manifest: %v", err)
 			}
