+++ pkg/model/utils.go
@@ -172,15 +172,6 @@ func dfs(s *Stack, svcName string, visited, stack map[string]bool) bool {
 	return false
 }
 
-// FileExistsAndNotDir checks if the file exists and its not a dir
-func FileExistsAndNotDir(filename string) bool {
-	info, err := os.Stat(filename)
-	if err != nil && os.IsNotExist(err) {
-		return false
-	}
-	return !info.IsDir()
-}
-
 func pathExistsAndDir(path string) bool {
 	info, err := os.Stat(path)
 	if err != nil && os.IsNotExist(err) {
